Introduce proxy.core.registration package to generalize proxy reg
The new proxy.core.registration.Registration provides the ProxyOptions
and ProxyMain subclasses specific to the proxy.
Next to it the Registrar class handles the registration.
As an implementation detail - so, as of now - it wraps the Config class'
related methods, but this may be changed.

As a result the proxy main loop in Main.java is simplified,
adding a new proxy does not require any change in it (in theory, of course).