ssh: buffer supports method chaining for putX; remove redundant bufferEnd
bufferEnd had the same value as size.
The size was in use, so it is kept with name 'limit'
in the way java.nio (Buffer, ByteBuffer) uses it: the amount of valid
data of the buffer - the first invalid index where data should not be read.

The basic tests are slightly enhanced.