ssh: load host key, add digest and prepare for KEXDH_INIT packet
The KeyExchange loads the host key, makes some preps for the
next - expected - SSH_MSG_KEXDH_INIT.
The Buffer now supports mpints via 'BigInteger getMpInt()'.

The TransportLayer gets the handlePacketsInLoop() method which is good
till the other side is in use. From that point a refactor is needed
to work in non-blocking mode.

The host key must be at ~/.config/jada/ssh_host_rsa_key
which may be RSA private key but the code doesn't check it as of now.

The SHAx digests are also added.
The diffie-hellman code exchange is partially added, at least
as dummy classes and with the group 1, 14, 16 and 18 constants.