ssh: partially rewritten to use Apache SSHD-Mina, partial KEXDH
The kexdh part is partially implemented, some parts are from SSHD-Mina.
As this part is complex and for the planned functionality reimplementing
the complete key exchange has no real advantage, the plan is to migrate
the related code to the mina's corresponding code.

I commit all changes I was able to figure out, but as I mentioned these
changes will be removed or rewritten to wrap the Apache code.

I use Bouncy castle as crypto provider, it's also registered in that way
via "Sec" class - it should be in SecurityUtils or ProviderFactory or so.

As a result the custom Buffer class is not needed, but the Apache
ByteArrayBuffer needs to be customized as Packet, just like earlier
the plain Buffer<T>.